+++ src/modules/AnalogClock/AnalogClock.tsx
@@ -1,16 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import ClockUI from './ClockUI';
+import { Interval } from '../types';
 
-type IntervalStatus = 'passed' | 'current' | 'upcoming';
 
-type Interval = {
-  startTime: string;
-  endTime: string;
-  icon: React.ReactElement;
-  color: string;
-  label: string;
-  status?: IntervalStatus;
-};
 
 type Options = {
   showIcons: boolean;
@@ -45,37 +37,30 @@ const AnalogClock: React.FC<AnalogClockProps> = ({
 
   useEffect(() => {
     const filterIntervals = (currentTime: Date) => {
-      const currentMinutes =
-        currentTime.getHours() * 60 + currentTime.getMinutes();
       const fiveMinutesAgo = new Date(currentTime.getTime() - 5 * 60000);
       const fortyFiveMinutesLater = new Date(
         currentTime.getTime() + 45 * 60000,
       );
 
       return intervals.reduce((acc: Interval[], interval: Interval) => {
-        const start = new Date(`1970-01-01T${interval.startTime}:00`);
-        const end = new Date(`1970-01-01T${interval.endTime}:00`);
-        const startMinutes = start.getHours() * 60 + start.getMinutes();
-        const endMinutes = end.getHours() * 60 + end.getMinutes();
-
-        if (startMinutes > currentMinutes + 45) return acc;
+        if (interval.startTime.getTime() > currentTime.getTime() + 45 * 60000)
+          return acc;
 
         let newInterval = { ...interval };
 
-        if (startMinutes < currentMinutes - 5) {
-          if (endMinutes <= currentMinutes - 5) return acc;
-          newInterval.startTime = fiveMinutesAgo.toTimeString().slice(0, 5);
+        if (interval.startTime.getTime() < currentTime.getTime() - 5 * 60000) {
+          if (interval.endTime.getTime() <= currentTime.getTime() - 5 * 60000)
+            return acc;
+          newInterval.startTime = fiveMinutesAgo;
         }
 
-        if (endMinutes > currentMinutes + 45) {
-          newInterval.endTime = fortyFiveMinutesLater
-            .toTimeString()
-            .slice(0, 5);
+        if (interval.endTime.getTime() > currentTime.getTime() + 45 * 60000) {
+          newInterval.endTime = fortyFiveMinutesLater;
         }
 
-        if (endMinutes <= currentMinutes) {
+        if (interval.endTime.getTime() <= currentTime.getTime()) {
           newInterval.status = 'passed';
-        } else if (startMinutes > currentMinutes) {
+        } else if (interval.startTime.getTime() > currentTime.getTime()) {
           newInterval.status = 'upcoming';
         } else {
           newInterval.status = 'current';
@@ -88,15 +73,11 @@ const AnalogClock: React.FC<AnalogClockProps> = ({
     setFilteredIntervals(filterIntervals(time));
   }, [time, intervals]);
 
-  const hours = time.getHours();
-  const minutes = time.getMinutes();
-  const seconds = time.getSeconds();
-
   return (
     <ClockUI
-      hours={hours}
-      minutes={minutes}
-      seconds={seconds}
+      hours={time.getHours()}
+      minutes={time.getMinutes()}
+      seconds={time.getSeconds()}
       intervals={filteredIntervals}
       showIcons={showIcons}
     />
