--- src/modules/CentralModule/types.ts
@@ -1,7 +1,0 @@
-export type Interval = {
-  startTime: string;
-  endTime: string;
-  icon: React.ReactElement | string;
-  color: string;
-  label: string;
-};
